refactor(node-compiler): FR-677 extract timeout wrapper to node_timeout module
Behavior-preserving split to reclaim headroom under the 450-line gate
(node_compiler 447 -> 397). Temporarily bump module-map budget 250 -> 260
to admit the new module; a follow-up FR reorganizes modules.
