+++ yamlgraph/node_compiler.py
@@ -4,10 +4,8 @@ Extracted from graph_loader.py to keep modules under 400 lines.
 Uses a registry pattern (FR-220) to dispatch node types to handlers.
 """
 
-import concurrent.futures
 import logging
 from collections.abc import Callable
-from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
@@ -27,6 +25,7 @@ from yamlgraph.node_factory import (
     create_subgraph_node,
     create_tool_call_node,
 )
+from yamlgraph.node_timeout import _maybe_wrap_timeout
 from yamlgraph.tools.agent import create_agent_node
 from yamlgraph.tools.nodes import create_tool_node
 from yamlgraph.tools.python_tool import create_python_node
@@ -91,56 +90,7 @@ def _parse_cache_field(raw: Any) -> CacheConfig | None:
     return None
 
 
-# Timeout wrapper (FR-069)
-
-
-def _maybe_wrap_timeout(
-    node_fn: Callable,
-    node_config: dict[str, Any],
-    node_name: str,
-) -> Callable:
-    """Wrap node function with ThreadPoolExecutor timeout if configured.
-
-    FR-069: Per-node timeout bounding. When timeout is set, the node
-    function is executed in a one-shot ThreadPoolExecutor. On
-    concurrent.futures.TimeoutError, a PipelineError with
-    error_type=TIMEOUT_ERROR is returned.
-
-    Args:
-        node_fn: The original node function
-        node_config: Node configuration dict (checked for 'timeout')
-        node_name: Name of the node (for error messages)
-
-    Returns:
-        Wrapped function if timeout is set, original function otherwise
-    """
-    timeout = node_config.get("timeout")
-    if timeout is None:
-        return node_fn
-
-    state_key = node_config.get("state_key", node_name)
-
-    def timed_fn(state: dict) -> dict:
-        pool = ThreadPoolExecutor(max_workers=1)
-        try:
-            return pool.submit(node_fn, state).result(timeout=timeout)
-        except concurrent.futures.TimeoutError as e:
-            from yamlgraph.models import PipelineError
-            from yamlgraph.models.schemas import ErrorType
-
-            pe = PipelineError.from_exception(
-                e, node=node_name, error_type=ErrorType.TIMEOUT_ERROR
-            )
-            return {
-                state_key: None,
-                "current_step": node_name,
-                "errors": [pe],
-            }
-        finally:
-            pool.shutdown(wait=False, cancel_futures=True)
-
-    timed_fn.__name__ = getattr(node_fn, "__name__", f"{node_name}_node")
-    return timed_fn
+# Timeout wrapper (FR-069): see yamlgraph.node_timeout._maybe_wrap_timeout
 
 
 # Node type handlers — one per node type
